Add support for Owon PDS7102T
Some hacking is required as the timebase is factured as 5, 10, 20 on all
the new scopes, not 5, 10, 25 as on the PDS5022S

IMHO, the timebase is not really what you want to know. The timebase is
the time it takes to render one div on the screen - but there are
between 100 and 625 samples per div.

For example: the time between each sample is the same, 2ns, for the 3
timebases 200ns, 500ns and 1us. Only the samples per div changes 100,
250, 500.