Unwrap sample dump
Make sure that the samples are dumped in chronological
order to the output file.

Description of samplecount1, samplecount2 and startoffset:

The header field "startoffset" points to the oldest sample
(more or less) and wrap around to the beginning when crossing
"samplecount2". All in all there are "samplecount1" valid
samples in the buffer.